LayoutTests/fast/js/mozilla/strict/12.14.1-expected.txt

PASS Function("'use strict'; try{}catch(eval){}") threw exception of type SyntaxError.
PASS !!Function("try{}catch(eval){}") is true
PASS true === true
PASS Function("'use strict'; try{}catch([eval]){}") threw exception of type SyntaxError.
FAIL !!Function("try{}catch([eval]){}") should be true. Threw exception SyntaxError: Parse error
PASS true === true
PASS Function("'use strict'; try{}catch({x:eval}){}") threw exception of type SyntaxError.
FAIL !!Function("try{}catch({x:eval}){}") should be true. Threw exception SyntaxError: Parse error
PASS true === true
PASS Function("'use strict'; try{}catch(arguments){}") threw exception of type SyntaxError.
PASS !!Function("try{}catch(arguments){}") is true
PASS true === true
PASS Function("'use strict'; try{}catch([arguments]){}") threw exception of type SyntaxError.
FAIL !!Function("try{}catch([arguments]){}") should be true. Threw exception SyntaxError: Parse error
PASS true === true
PASS Function("'use strict'; try{}catch({x:arguments}){}") threw exception of type SyntaxError.
FAIL !!Function("try{}catch({x:arguments}){}") should be true. Threw exception SyntaxError: Parse error
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
